# Speed up directory loading a bit