--- a/romsel_dsimenutheme/arm9/source/fileBrowse.h
+++ b/romsel_dsimenutheme/arm9/source/fileBrowse.h
@@ -29,6 +29,6 @@
 
 bool extension(const std::string_view filename, const std::vector<std::string_view> extensions);
 
-std::string browseForFile(const std::vector<std::string> extensionList);
+std::string browseForFile(const std::vector<std::string_view> extensionList);
 
 #endif //FILE_BROWSE_H