--- a/romsel_r4theme/arm9/source/fileBrowse.h
+++ b/romsel_r4theme/arm9/source/fileBrowse.h
@@ -29,7 +29,7 @@
 
 bool extension(const std::string_view filename, const std::vector<std::string_view> extensions);
 
-std::string browseForFile(const std::vector<std::string> extensionList);
+std::string browseForFile(const std::vector<std::string_view> extensionList);