<!-- ##### REMEMBER TO ALWAYS TEST YOUR PR! -->
#### What's changed?

- Speeds up directory loading a bit, I think it cuts the directory load time about in half, but game info loading is only slightly improved
   - A folder with 300 copies of BOOT.NDS goes from 8s to 4.5s for me
   - My DS ROM folder goes from 2s to 1.5s for me
- The main slow part now is checking if files are hidden, with hidden files shown it's only about 1.5s for any folder with 40+ items for me
- Also updates the R4 theme extension() function

#### Where have you tested it?

- DSi (K) from Unlaunch

***

#### Pull Request status
- [x] This PR has been tested using the latest version of devkitARM and libnds.